Reorganize types
